Revive outdated method Lab.recreate_student_invitations.
Used to be in Course.
