+++ course.py
@@ -15,7 +15,6 @@ from pathlib import Path
 from typing import Iterable
 
 import atomicwrites
-import dateutil.parser
 import gitlab
 import ldap
 import more_itertools
@@ -604,38 +603,6 @@ class Course:
             self.logger.debug(f"deleting obsolete invitation of {email}")
             with gitlab_.tools.exist_ok():
                 gitlab_.tools.delete(self.gl, self.graders_group.lazy, email)
-
-    def recreate_student_invitations(self, keep_after=None):
-        """
-        Recreate student invitations to groups on Chalmers GitLab.
-        Since GitLab does not expose an API for resending invitations,
-        we delete and recreate invitations instead.
-
-        If the date keep_after is given (instance of datetime.datetime),
-        only those invitations are recreated that have been created before the given date.
-        """
-        earlier_than = "" if keep_after is None else f" earlier than {keep_after}"
-        self.logger.info(f"recreating student invitations{earlier_than}.")
-
-        for group_id in self.groups:
-            entity = self.group(group_id).lazy
-            entity_name = f"{self.config.group.name.print(group_id)} on GitLab"
-            for invitation in gitlab_.tools.invitation_dict(self.gl, entity).values():
-                created_at = dateutil.parser.parse(invitation["created_at"])
-                if keep_after is None or created_at < keep_after:
-                    email = invitation["invite_email"]
-                    self.logger.info(
-                        f"Recreating invitation from {created_at} of {email} to {entity_name}."
-                    )
-                    with gitlab_.tools.exist_ok():
-                        gitlab_.tools.invitation_delete(self.gl, entity, email)
-                    try:
-                        with gitlab_.tools.exist_ok():
-                            gitlab_.tools.invitation_create(
-                                self.gl, entity, email, gitlab.const.DEVELOPER_ACCESS
-                            )
-                    except gitlab.exceptions.GitlabCreateError as e:
-                        self.logger.error(str(e))
 
     def student_members(self, cached_entity):
         """
